Created speed and color effect script
Color script will assign a particle color from the collision object. 

Speed will try to scale particle count to collision speed. It needs more work as its not having the desired effect. Mainly there is not a massive visual difference. Which will need to be solved by adjusting how the particles spawn.

There also seems to be an issue were the effect spawn on a different location from the collision spot.